chore: extract centeredBox and scene-depth helpers shared by scene pages
Deduplicate between DynamicScenePage and ComicBookReader:
- centeredBox: full-viewport flex container for loading/error/empty states
  (src/utils/pageLayout.js)
- computeMaxZ / computeScrollDepth: scene depth math
  (src/utils/sceneDepth.js)
