+++ src/pages/ComicBookReader.jsx
@@ -8,6 +8,8 @@ import { useComicBook } from '../hooks/useComicBook';
 import { useZScroll } from '../hooks/useZScroll';
 import { useThemeTriggers } from '../hooks/useThemeTriggers';
 import { getLayerUrl } from '../services/gcsStorage';
+import { centeredBox } from '../utils/pageLayout';
+import { computeMaxZ, computeScrollDepth } from '../utils/sceneDepth';
 
 /**
  * ComicBookReader - Read-only viewer for published comic books loaded from GCS.
@@ -90,12 +92,7 @@ export function ComicBookReader() {
     mouseInfluence = { x: 50, y: 30 },
   } = sceneConfig;
 
-  const maxZ = useMemo(() => {
-    const layerZs = layers.map((l) => (l.position || [0, 0, 0])[2]);
-    const objectZs = objects.map((o) => (o.position || [0, 0, 0])[2]);
-    const allZs = [...layerZs, ...objectZs];
-    return allZs.length ? Math.max(...allZs) : 0;
-  }, [layers, objects]);
+  const maxZ = useMemo(() => computeMaxZ(layers, objects), [layers, objects]);
 
   const slides = useMemo(
     () =>
@@ -107,9 +104,10 @@ export function ComicBookReader() {
     [layers],
   );
 
-  const scrollDepth = useMemo(() => {
-    return (maxZ || 500) + perspective;
-  }, [maxZ, perspective]);
+  const scrollDepth = useMemo(
+    () => computeScrollDepth(maxZ, perspective),
+    [maxZ, perspective],
+  );
 
   const { scrollZ, containerRef } = useZScroll({
     slides,
@@ -139,20 +137,9 @@ export function ComicBookReader() {
     }
   }, [activeThemeName, currentScene, setTheme]);
 
-  const centeredBox = {
-    width: '100%',
-    height: '100vh',
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    background: theme.colors.backgroundGradient,
-    fontFamily: theme.typography.fontBody,
-  };
-
   if (loading) {
     return (
-      <div style={centeredBox}>
+      <div style={centeredBox(theme)}>
         <div style={{ color: theme.colors.textMuted, fontSize: '14px', letterSpacing: '2px' }}>
           Loading…
         </div>
@@ -162,7 +149,7 @@ export function ComicBookReader() {
 
   if (error) {
     return (
-      <div style={centeredBox}>
+      <div style={centeredBox(theme)}>
         <div
           style={{
             color: theme.colors.primary,
@@ -182,7 +169,7 @@ export function ComicBookReader() {
 
   if (!currentScene) {
     return (
-      <div style={centeredBox}>
+      <div style={centeredBox(theme)}>
         <div style={{ color: theme.colors.textMuted, fontSize: '14px', letterSpacing: '2px' }}>
           No scenes found
         </div>
